Extract MomentAboutMean logic
Making the code DRYer one stats Op at a time

Plus, if (for whatever reason) a user needed to find the 5th moment, or
the 6th moment, or the 28th moment, they can now do it.
