+++ imagej/imagej-ops2/src/main/java/net/imagej/ops2/stats/IterableMoment4AboutMean.java
@@ -29,10 +29,8 @@
 
 package net.imagej.ops2.stats;
 
-import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.type.numeric.RealType;
 
-import org.scijava.Priority;
 import org.scijava.ops.OpDependency;
 import org.scijava.ops.core.Op;
 import org.scijava.ops.function.Computers;
@@ -41,25 +39,41 @@ import org.scijava.plugin.Plugin;
 import org.scijava.struct.ItemIO;
 
 /**
- * {@link Op} to calculate the {@code stats.moment3AboutMean} using
+ * {@link Op} to calculate the {@code stats.moment4AboutMean} using
  * {@code stats.mean} and {@code stats.size}.
  * 
- * @author Gabriel Selzer
+ * @author Daniel Seebacher (University of Konstanz)
+ * @author Christian Dietz (University of Konstanz)
  * @param <I>
  *            input type
  * @param <O>
  *            output type
  */
-@Plugin(type = Op.class, name = "stats.moment3AboutMean", priority = Priority.HIGH)
+@Plugin(type = Op.class, name = "stats.moment4AboutMean")
 @Parameter(key = "iterableInput")
-@Parameter(key = "moment3AboutMean", itemIO = ItemIO.BOTH)
-public class DefaultMoment3AboutMean<I extends RealType<I>, O extends RealType<O>> implements Computers.Arity1<RandomAccessibleInterval<I>, O> {
+@Parameter(key = "moment4AboutMean", itemIO = ItemIO.BOTH)
+public class IterableMoment4AboutMean<I extends RealType<I>, O extends RealType<O>> implements Computers.Arity1<Iterable<I>, O> {
 
-	@OpDependency(name = "stats.momentNAboutMean")
-	private Computers.Arity2<RandomAccessibleInterval<I>, Integer, O> momentComputer;
+	@OpDependency(name = "stats.mean")
+	private Computers.Arity1<Iterable<I>, O> meanComputer;
+	@OpDependency(name = "stats.size")
+	private Computers.Arity1<Iterable<I>, O> sizeComputer;
 
 	@Override
-	public void compute(final RandomAccessibleInterval<I> input, final O output) {
-		momentComputer.compute(input, 3, output);
+	public void compute(final Iterable<I> input, final O output) {
+
+		final O mean = output.createVariable();
+		meanComputer.compute(input, mean);
+		final O size = output.createVariable();
+		sizeComputer.compute(input, size);
+
+		double res = 0;
+		final double m = mean.getRealDouble();
+		for (final I in : input) {
+			final double val = in.getRealDouble() - m;
+			res += val * val * val * val;
+		}
+
+		output.setReal(res / size.getRealDouble());
 	}
 }
